+++ src/documentation/pages/templates/error-500/examples/index.vue
@@ -1,6 +1,6 @@
 <template lang="pug">
   div
-    div(v-if="selectedTheme === 'lumen' || selectedTheme === 'portal'")
+    div(v-if="['lumen', 'lumenrebrand24', 'portal', 'portalrebrand24'].includes(selectedTheme)")
       BaseExample
 
     div(v-if="selectedTheme === 'centurylink'")