+++ src/website/assets/scripts/globalConfigs.js
@@ -1,1 +1,1 @@
-window.chiCurrentVersion="5.64.0";
+window.chiCurrentVersion="5.69.0";